Moved sitenow:multisite:delete to uiowa:multisite:delete.
